requirements.txt
schema==0.7.5
tqdm==4.64.0
SQLAlchemy~=1.4.41
fastapi==0.85.0
uvicorn==0.18.3
Faker==13.12.1
pymysql==1.0.2
pydantic~=1.10.2
python-multipart==0.0.5
Pillow==9.2.0
websockets==10.3
psycopg2-binary==2.9.3
passlib==1.7.4
PyJWT==2.5.0
